+++ shv/__init__.py
@@ -59,7 +59,7 @@ from .shvclient import SHVClient
 from .shvdevice import SHVDevice
 from .shvmethods import SHVMethods
 from .shvvalueclient import SHVValueClient
-from .shvversion import SHV_VERSION_MAJOR, SHV_VERSION_MINOR
+from .shvversion import SHV_VERSION, SHV_VERSION_MAJOR, SHV_VERSION_MINOR
 from .value import (
     SHVBool,
     SHVBoolType,
@@ -93,6 +93,7 @@ from .value import (
 )
 
 __all__ = [
+    "SHV_VERSION",
     "SHV_VERSION_MAJOR",
     "SHV_VERSION_MINOR",
     "VERSION",